requirements.txt
nemoguardrails>=0.11.1
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
langsmith>=0.0.69
pydantic>=2.5.0
faiss-cpu>=1.7.4
langchain-community>=0.0.13

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0 